Refactor countAllComments test
