+++ __tests__/countAllComments.test.js
@@ -2,12 +2,22 @@
  * @jest-environment jsdom
  */
 
-import { displayMeals, postComment, getComment, getAllComments, countAllComments } from '../__mocks__/countAllCommentsMock.js';
- 
-describe('count all comment test', () => {
-  test('comment node 2 should have length of 12', () => {
-    countAllComments();
+import { countAllComments } from '../src/modules/getAllMeals.js';
+import '@testing-library/jest-dom';
 
-    //expect(countAllComments()).toBe(12);
+describe('count all comment test', () => {
+  test('comment counter should have length of 5', () => {
+    document.body.innerHTML = `
+    <span class="comment-counter"></span>
+    <ul class="user-comments">
+    <li class="comment-item"></li>
+    <li class="comment-item"></li>
+    <li class="comment-item"></li>
+    <li class="comment-item"></li>
+    <li class="comment-item"></li>
+    </ul>`;
+    countAllComments();
+    const length = document.querySelector('.comment-counter').textContent;
+    expect(length).toBe(' (5)');
   });
 });
